pyproject.toml
[project]
name = "hormuz-supply-chain"
version = "0.1.0"
description = "Hormuz supply chain data pipeline and explorer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.0.0",
    "pandas>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "tqdm>=4.65.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]
